move helper functions into separate include
